Fix for: 'Validate' button broken if feed URL contains ampersand.
The 'Validate' button is broken if the feed URL contains an ampersand.
This is because the validateFeed javascript function submits the 'url'
variable unescaped in a GET request, leading to the $sourcefeed variable
in the Perl part only containing the first part of the URL up to the
ampersand.

Wrapped URL in escape() function to avoid this problem.